+++ src/application/tests/testResponseBodyContent.test.ts
@@ -1,12 +1,17 @@
-import { getPostmanMappedOperation } from '../../../__tests__/testUtils/getPostmanMappedOperation'
+import {
+  getPostmanMappedListArrayOperation,
+  getPostmanMappedOperation
+} from '../../../__tests__/testUtils/getPostmanMappedOperation'
 import { testResponseBodyContent } from '../../application'
 import { PostmanMappedOperation } from '../../postman'
 
 describe('testResponseBodyContent', () => {
   let pmOperation: PostmanMappedOperation
+  let pmArrayOperation: PostmanMappedOperation
 
   beforeEach(async () => {
     pmOperation = await getPostmanMappedOperation()
+    pmArrayOperation = await getPostmanMappedListArrayOperation()
   })
 
   afterEach(() => {
@@ -129,6 +134,149 @@ describe('testResponseBodyContent', () => {
     ]
     pmOperation = testResponseBodyContent(contentTests, pmOperation)
     const pmTest = pmOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for root array & result has minimum length value', async () => {
+    const contentTests = [
+      {
+        key: '.',
+        minLength: 1
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmArrayOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for  root array & result has maximum length value', async () => {
+    const contentTests = [
+      {
+        key: '.',
+        maxLength: 10
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmArrayOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for string for root array property check & string value', async () => {
+    const contentTests = [
+      {
+        key: '[0].company_name',
+        value: 'Spacex'
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for string for root array property check & boolean value', async () => {
+    const contentTests = [
+      {
+        key: '[0].data.company_name',
+        value: true
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for root array property check & number value', async () => {
+    const contentTests = [
+      {
+        key: '[0].data.monetary_amount',
+        value: 75000
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for root array property check without value', async () => {
+    const contentTests = [
+      {
+        key: '[0].data.monetary_amount'
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for root array property check & pm variable value', async () => {
+    const contentTests = [
+      {
+        key: '[0].data.monetary_amount',
+        value: '{{postman_env_variable}}'
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for root array property check & result contains value', async () => {
+    const contentTests = [
+      {
+        key: '[0].data.name',
+        contains: 'Musk'
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for root array property check & result contains pm variable value', async () => {
+    const contentTests = [
+      {
+        key: '[0].data.monetary_amount',
+        contains: '{{postman_env_variable}}'
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for root array property check & result has length value', async () => {
+    const contentTests = [
+      {
+        key: '[0].data.description',
+        length: 9
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for root array property check & result has minimum length value', async () => {
+    const contentTests = [
+      {
+        key: '[0].data',
+        minLength: 1
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test forroot array property check & result has maximum length value', async () => {
+    const contentTests = [
+      {
+        key: '[0].data',
+        maxLength: 10
+      }
+    ]
+    pmArrayOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmArrayOperation.getTests()
     expect(pmTest.script.exec).toMatchSnapshot()
   })
 })